Fix formatting again.

+++ MagicKeys/Core/Mouse/Convert.cs
@@ -1,17 +1,18 @@
 using System;
-using System.Runtime.InteropServices;
-using System.Windows.Forms;
+
 namespace MagicKeys
 {
 public partial class MagicKeys
 {
 public static int MouseConvertX(int X)
 {
-return Convert.ToInt32(MXDC*(decimal)X);
+return Convert.ToInt32(MXDC* X);
 }
+
 public static int MouseConvertY(int Y)
 {
-return Convert.ToInt32(MYDC*(Decimal)Y);
+return Convert.ToInt32(MYDC* Y);
 }
+
 }
 }
